ACPI / irq: Add acpi_register_partitioned_percpu_gsi()
A percpu interrupt can occur on any CPU, or multiple CPUs in parallel.
These are used when the source of the interrupt is the CPU itself,
or a device that is integrated with the CPU.

When a device is integrated with a group of CPUs, this device can't
trigger the interrupt on a CPU outside that group. There may be
multiple devices, which between them can trigger the percpu interrupt
on any CPU.

Percpu interrupts can only be registered once. If there are multiple
devices using the same percpu interrupt, the driver must be aware
that its percpu interrupt could occur on a CPU whose device has not yet
been probed.

Add an acpi helper to register partitioned percpu interrupts. These
take a processor container to describe which CPUs the interrupt can
occur on. The returend irq can be requested once per partition,
allowing drivers to treat these devices independently.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>
(cherry picked from commit 9579f90f164450fa447b1f11f372f83911fc3857)
Signed-off-by: kobak <kobak@nvidia.com>