ACPI / PPTT: Add a helper to fill a cpumask from a cache_id
MPAM identifies CPUs by the cache_id in the PPTT cache structure.

The driver needs to know which CPUs are associated with the cache,
the CPUs may not all be online, so cacheinfo does not have the
information.

Add a helper to pull this information out of the PPTT.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>
(cherry picked from commit 986b6d7c9aae160db9625003808edece0b9c5207)
Signed-off-by: kobak <kobak@nvidia.com>