arm_mpam: resctrl: Allow resctrl to allocate monitors
When resctrl wants to read a domain's 'QOS_L3_OCCUP', it needs
to allocate a monitor on the corresponding resource. Monitors are
allocated by class instead of component because any per-component
user needs to have pre-emption disabled to avoid being migrated to
another CPU.

Add helpers to do this.

The MBM events depend on having their monitors allocated at init
time so that they can be left running. The value USE_RMID_IDX
is out of range for a monitor, and is used to indicate this
behaviour.

This patch temporarily creates resctrl_mon_ctx_waiters as the
resctrl version can't be selected until it will link. This gets
removed in a later patch.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>
Signed-off-by: KobaK <kobak@nvidia.com>