arm64: mpam: Add helpers to change a tasks and cpu mpam partid/pmg values
Care must be taken when modifying the partid and pmg of a task, as
writing these values may race with the task being scheduled in, and
reading the modified values.

Add helpers to set the task properties, and the cpu default value,
and add the plumbing to the mpam driver that lets resctrl use them.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>
(cherry picked from commit 75a171443a2733e56065bed7c1372d372d2a4792)
Signed-off-by: kobak <kobak@nvidia.com>