cacheinfo: Expose the code to generate a cache-id from a device_node
The MPAM driver identifies caches by id for use with resctrl. It
needs to know the cache-id when probeing, but the value isn't set
in cacheinfo until device_initcall().

Expose the code that generates the cache-id. The parts of the MPAM
driver that run early can use this to set up the resctrl structures.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>
(cherry picked from commit a171c1e14d008851957278e230f4f99597680fa6)
Signed-off-by: kobak <kobak@nvidia.com>