arm_mpam: Allow the maximum partid to be overridden from the command line
MPAMs bandwidth monitors are only available via resctrl if there are
enough monitors for each combination of partid and pmg to have one.

As it is unlikely anyone built that many monitors, allow the
maximum partid the system will use to be set from the kernel
command-line.

With this, it should be possible to bandwidth monitors to be
enabled by reducing the number of partid in use.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>
(cherry picked from commit 894a5221d6f836f09ca61f23a22ed9b9981fa2da)
Signed-off-by: kobak <kobak@nvidia.com>