fs/resctrl: Add helper to query if counters are free running
Resctrl exposes the cache opccupancy and bandwidth counters to user space via
files. Each control and monitor group has a copy of these files, and can
read the coutner by reading the file.

MPAM needs to allocate a hardware monitor to do the counting work, and there
may not be enough for every control and monitor group to have one.

Indicate this support via a resctrl_arch_event_is_free_running() helper.
This indicates whether the event counts even when no architecture context
has been allocated to read it.

This is always true for all events on x86.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>
Signed-off-by: KobaK <kobak@nvidia.com>