x86/resctrl: Allow resctrl_arch_mon_event_config_write() to return an error
resctrl_arch_mon_event_config_write() writes a bitmap of events provided
by user-space into the configuration register for the monitors.

This assumes that all architectures support all the features each bit
corresponds to.

MPAM can filter monitors based on read, write, or both, but there are
many more options in the existing bitmap. To allow this interface to
work for machines with MPAM, allow the architecture helper to return
an error if an incompatible bitmap is set.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>
Signed-off-by: KobaK <kobak@nvidia.com>