arm_mpam: Add cranky debug for reading CSU hardware values
MPAM affects performance, which makes it hard to debug. (is that the
performance I was expecting?). Another problem is when multiple MSC
get combined into a single component, how do I know one of them isn't
being missed out - or isn't counting a value.

As a debug aid, add hw_ files to debugfs that read the counter hardware
directly.

This is not for upstream.

This interface has no visibility of what the monitors were allocated
for, and they may be re-allocated and re-programmed between accesses.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>
Signed-off-by: KobaK <kobak@nvidia.com>